requirements.txt
# Real-Time Fraud Detection System
# Python 3.9+ required
# Java 11 required (for Spark)

# Core Streaming & Messaging
kafka-python==2.0.2

# Spark (requires Java 11 to be installed separately)
pyspark==3.5.0

# Database Drivers
psycopg2-binary==2.9.9
redis==5.0.1

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Processing
numpy==1.26.2
pandas==2.1.3

# HTTP & Utilities
requests==2.31.0
python-dotenv==1.0.0

# Optional: For Jupyter analysis notebooks
# jupyter==1.0.0
# matplotlib==3.8.2
# seaborn==0.13.0
